Allow better tree shaking:
Move helpers to separate files
Add sideEffects false to package.json
