+++ src/__tests__/createBreakpointQueryObject.test.ts
@@ -1,18 +1,12 @@
-import {
-  minWidth,
-  maxWidth,
-  minMaxWidth,
-  createBreakpointQueryObject,
-} from '../media-utils';
+import { createBreakpointQueryObject } from '../helpers/createBreakpointQueryObject';
 
 describe('createBreakpointQueryObject', () => {
   it('creates the correct query for sm, md, lg', () => {
-    const breakpointUnit = 'px';
     const breakpoints = { sm: 10, md: 20, lg: 30 };
     const media = {
-      sm: maxWidth(breakpoints.md, breakpointUnit),
-      md: minMaxWidth(breakpoints.md, breakpoints.lg, breakpointUnit),
-      lg: minWidth(breakpoints.lg, breakpointUnit),
+      sm: '(max-width: 20px)',
+      md: '(min-width: 20px) and (max-width: 30px)',
+      lg: '(min-width: 30px)',
     };
 
     expect(createBreakpointQueryObject(breakpoints, 'px')).toMatchObject(media);
