+++ src/helpers/convertScreenWidth.ts
@@ -1,0 +1,8 @@
+import type { BreakpointUnit } from '../breakpoints';
+
+export function convertScreenWidth(
+  screenWidth: number,
+  breakpointUnit: BreakpointUnit,
+): number {
+  return breakpointUnit === 'em' ? screenWidth / 16 : screenWidth;
+}
