+++ src/helpers/messages.ts
@@ -5,4 +5,4 @@ export const ERRORS = {
   NO_BREAKPOINTS: `react-breakpoints requires you to pass an object of breakpoints, ${DOCS}`,
   NOT_OBJECT: `react-breakpoints requires you to pass an object of breakpoints, ${DOCS}`,
   EMPTY_OBJECT: `react-breakpoints requires you to pass an object of breakpoints, ${DOCS}`,
-};
+} as const;
