+++ src/helpers/sortBreakpoints.ts
@@ -1,0 +1,8 @@
+import type { BreakpointMap, SortedBreakpoints } from '../breakpoints';
+
+/**
+ * Sort from largest to smallest
+ */
+export function sortBreakpoints(breakpoints: BreakpointMap): SortedBreakpoints {
+  return Object.entries(breakpoints).sort((a, b) => b[1] - a[1]);
+}
