+++ src/withBreakpoints.tsx
@@ -1,7 +1,7 @@
 import { ComponentType, useContext } from 'react';
 import hoistStatics from 'hoist-non-react-statics';
 
-import { BreakpointKey, BreakpointsProps } from './breakpoints';
+import type { BreakpointKey, BreakpointsProps } from './breakpoints';
 import { BreakpointsContext } from './BreakpointsContext';
 
 /**
